Added spans and moved split
